balkanize capturing and logging worker status information
use RingBuffer everywhere
