+++ src/Ilúvatar/iluvatar_library/src/transaction.rs
@@ -11,7 +11,6 @@ lazy_static::lazy_static! {
   pub static ref STARTUP_TID: TransactionId = "Startup".to_string();
   pub static ref TEST_TID: TransactionId = "TestTest".to_string();
   pub static ref LOAD_BALANCER_TID: TransactionId = "LoadBalancer".to_string();
-  pub static ref STATUS_WORKER_TID: TransactionId = "Status".to_string();
   pub static ref HEALTH_TID: TransactionId = "HealthCheck".to_string();
   pub static ref LEAST_LOADED_TID: TransactionId = "LeastLoadedLB".to_string();
   pub static ref SIMULATION_START_TID: TransactionId = "SimulationStartup".to_string();
