+++ src/Ilúvatar/iluvatar_worker_library/src/services/mod.rs
@@ -4,5 +4,4 @@ pub mod invocation;
 pub mod network;
 pub mod registration;
 pub mod resources;
-pub mod status;
 pub mod worker_health;
